% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{rk.ANOVA-package}
\alias{rk.ANOVA-package}
\title{The rk.ANOVA Package}
\description{
RKWard GUI to conduct ANOVAs (using the ez package),
      pairwise t-Tests and plot interactions..
}
\details{
\tabular{ll}{
Package: \tab rk.ANOVA\cr
Type: \tab Package\cr
Version: \tab 0.01-18\cr
Date: \tab 2014-09-14\cr
Depends: \tab  rkward (>= 0.5.6)\cr
Enhances: \tab rkward\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
URL: \tab http://rkward.sf.net\cr
}

RKWard GUI to conduct ANOVAs (using the ez package),
      pairwise t-Tests and plot interactions.
}
\author{
Meik Michalke
}
\keyword{package}

