% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rk.ANOVA-package.R
\docType{package}
\name{rk.ANOVA-package}
\alias{rk.ANOVA-package}
\title{The rk.ANOVA Package}
\description{
RKWard GUI to conduct ANOVAs (using the ez package),
      pairwise t-Tests and plot interactions..
}
\details{
\tabular{ll}{
Package: \tab rk.ANOVA\cr
Type: \tab Package\cr
Version: \tab 0.01-19\cr
Date: \tab 2015-03-13\cr
Depends: \tab rkward (>= 0.6.0)\cr
Enhances: \tab rkward\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
URL: \tab http://rkward.sf.net\cr
}

RKWard GUI to conduct ANOVAs (using the ez package),
      pairwise t-Tests and plot interactions.
}
\author{
Meik Michalke
}
\keyword{package}

