\docType{package}
\name{rk.ANOVA-package}
\alias{rk.ANOVA}
\alias{rk.ANOVA-package}
\title{The rk.ANOVA Package}
\description{
  RKWard GUI to conduct ANOVAs (using the ez package),
  pairwise t-Tests and plot interactions..
}
\details{
  \tabular{ll}{ Package: \tab rk.ANOVA\cr Type: \tab
  Package\cr Version: \tab 0.01-15\cr Date: \tab
  2013-02-21\cr Depends: \tab rkward (>= 0.5.6)\cr
  Enhances: \tab rkward\cr License: \tab GPL (>= 3)\cr
  LazyLoad: \tab yes\cr URL: \tab http://rkward.sf.net\cr }

  RKWard GUI to conduct ANOVAs (using the ez package),
  pairwise t-Tests and plot interactions.
}
\author{
  Meik Michalke \email{meik.michalke@hhu.de}
}
\keyword{package}

